testapp/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
opencv-python>=4.5.0
scikit-image>=0.18.0
pyyaml>=5.4.0
tqdm>=4.60.0
scipy>=1.7.0